refactor(board): extract shared test utils
Dedupe three verbatim makeBoardSet factories into board-sets/test-utils,
and the board fixture + render wrapper shared by the viewer and keyboard
tests into a root test-utils. Rename storage/test-helpers to test-utils
so colocated helpers share one name.
