+++ src/features/board/keyboard/use-board-keyboard.test.tsx
@@ -1,35 +1,7 @@
 import { stubAudio, stubSpeech } from "@shared/testing/device-output";
-import { AppProviders } from "@shared/providers/app-providers";
-import type { OBFBoard } from "@shayc/open-board-format";
-import { MemoryRouter } from "react-router";
 import { beforeEach, describe, expect, test, vi } from "vitest";
-import { render } from "vitest-browser-react";
 import { userEvent } from "vitest/browser";
-import { BoardViewer } from "../board-viewer";
-import { obfToBoard } from "../obf/obf-to-board";
-
-const TWO_TILE_BOARD: OBFBoard = {
-  format: "open-board-0.1",
-  id: "test-board",
-  locale: "en",
-  buttons: [
-    { id: "btn-1", label: "hello" },
-    { id: "btn-2", label: "world" },
-  ],
-  grid: { rows: 1, columns: 2, order: [["btn-1", "btn-2"]] },
-};
-
-async function renderBoard() {
-  return render(
-    <MemoryRouter>
-      <AppProviders>
-        <div style={{ height: "100vh" }}>
-          <BoardViewer board={obfToBoard(TWO_TILE_BOARD)} />
-        </div>
-      </AppProviders>
-    </MemoryRouter>,
-  );
-}
+import { renderBoardViewer, TWO_TILE_BOARD } from "../test-utils";
 
 describe("board keyboard shortcuts", () => {
   let speech: ReturnType<typeof stubSpeech>;
@@ -40,7 +12,7 @@ describe("board keyboard shortcuts", () => {
   });
 
   test("Backspace from a focused tile removes the last message part", async () => {
-    const screen = await renderBoard();
+    const screen = await renderBoardViewer(TWO_TILE_BOARD);
     await screen.getByRole("button", { name: "hello" }).click();
     await screen.getByRole("button", { name: "world" }).click();
 
@@ -59,7 +31,7 @@ describe("board keyboard shortcuts", () => {
   });
 
   test("Backspace works board-wide — even from a non-grid button", async () => {
-    const screen = await renderBoard();
+    const screen = await renderBoardViewer(TWO_TILE_BOARD);
     await screen.getByRole("button", { name: "hello" }).click();
     await screen.getByRole("button", { name: "world" }).click();
 
@@ -75,7 +47,7 @@ describe("board keyboard shortcuts", () => {
   });
 
   test("⌘+Enter speaks the message", async () => {
-    const screen = await renderBoard();
+    const screen = await renderBoardViewer(TWO_TILE_BOARD);
     await screen.getByRole("button", { name: "hello" }).click();
     await screen.getByRole("button", { name: "world" }).click();
