+++ src/features/board/testing.ts
@@ -1,1 +1,1 @@
-export { resetBoardsDB, seedBoardSets } from "./storage/test-helpers";
+export { resetBoardsDB, seedBoardSets } from "./storage/test-utils";
